pyproject.toml
[tool.poetry]
name = "narrative-generation"
version = "0.1.0"
description = ""
authors = ["Axel Vanraes <axel.vanraes@tiro.health>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-liquid = "^1.12.1"
fhirpathpy = "^1.1.1"
streamlit = "^1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
watchdog = "^5.0.2"
pytest = "^8.3.2"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
